Make DAAP server push out changes in metadata (using only +/- deltas, not resending everything).
This is getting kind of cool. Remove a file from your MythMusic /directory,
watch the metadata sweeper see it disappear, take it off the master list,
send a signal to all the plugins, and then you'll see it disappear from an
iTunes list.

All a bit fragile at the moment, and playlists are hosed.


git-svn-id: http://svn.mythtv.org/trunk@2940 4b91aca1-2a1f-0410-a73a-a78f2df169e8