Initial optimization of the new scheduler algorithms.  The bigger optimization will come next.
Have the database order the programs so that MarkOverlaps can be O(1)
instead of O(2).

Merge MarkOverlaps and MarkTooManys into AddNewRecords to avoid extra
passes that can be done together.

Have comp_recstart order the programs so that PruneRedundants can be
O(1) instead of O(2).

Merge MarkConflicts into PruneRedundants to avoid extra passes that
can be done together.


git-svn-id: http://svn.mythtv.org/trunk@3312 4b91aca1-2a1f-0410-a73a-a78f2df169e8