Add ability to dynamically load and resolve parts of libopendaap.
This will allow you to play audio content from within the mfd that is being
served by iTunes (ie. bi-directional music sharing)

As libopendaap was only released yesterday, you need to apply a patch
(included) to avoid an annoying segfault. Details are in the main README and
in ./mfd/plugins/daapclient/README.

Dynamic updates work for items (ie. files to play), but playlists are a bit
messed up at the moment.


git-svn-id: http://svn.mythtv.org/trunk@3328 4b91aca1-2a1f-0410-a73a-a78f2df169e8