Lots and lots of cosmetic changes; rewording some things to make it clearer (hopefully) what's going on.
Move important information closer to the front of the help text in case it's
cut off due to overscan.

Change some entries so that they're not the equivalent of
"Yes, I don't want to do something."

Re-format some code so that it's following the same standard - multi-line
text lines up under the "Q" in QObject.

NOTE:  Translators, there are a number of changes that will impact the text
strings.


git-svn-id: http://svn.mythtv.org/trunk@3894 4b91aca1-2a1f-0410-a73a-a78f2df169e8