Now with ProgramInfo goodness.
The root cause of the crash when doing an override has been fixed.  Full info is once again available.
The problem where the info from the record table is always used instead of info from the Programinfo has been fixed.

A side effect of this is that the title doesn't show up with it's search type anymore.  I plan to add a new element to the map so this info can be incorporated into the UI.


git-svn-id: http://svn.mythtv.org/trunk@3912 4b91aca1-2a1f-0410-a73a-a78f2df169e8