Added support for a default view.  This allows us to have one menu entry for MythVideo instead of one for each view type.  The jump point "MythVideo" will take you to your default view.
The metadata object can now provide a QImage of it's cover image if one exists.  This allows us to load the image once and reference when ever we have the metadata around for the video.

The video gallery now keeps a copy of the metadata for each video around.

The video gallery drawing code has been modified to use the copy of the cover art from the metadata object instead of loading it every time.  This should speed up gallery drawing, though it could still use some optimization.

Refactored all three views to use a common base class that deals with common tasks such as fetching data from the db, parental controls, launching the video player, common menu options etc.

All three views now have a MENU menu and an INFO menu.  The MENU menu allows you to jump to different views, and in the case of the Gallery different styles of the view.  The INFO menu contains, at a minimum "Play this Video", and "View full plot".

Removed a ton of tabs.

Fixed some "if(" to "if (".

Fixed several brackets and indentations.


git-svn-id: http://svn.mythtv.org/trunk@4083 4b91aca1-2a1f-0410-a73a-a78f2df169e8