AVFormatDecoder now properly handles files with multiple audio tracks with the stream (such as perfect DVD rips).  It *should* select the proper default audio track for you automaticly is you find it's making a lot of mistakes let me know, I may have made some bad assumptions.
Two new key bindings have been added for the TV Playback. The NEXTAUDIO (+) and PREVAUDIO (-) allow you to cycle through the various audio tracks within the stream (in case you really do want to hear the directors blathering on for example).

Hopefully this will clear up one of the major stumbling blocks for using the Internal player in MythVideo.


git-svn-id: http://svn.mythtv.org/trunk@4281 4b91aca1-2a1f-0410-a73a-a78f2df169e8