Added Advanced search. This is a user interface to the kPowerSearch capability added by gigem for arbitrary FROM and WHERE clause SQL to be embedded within a kPowerSearch record rule.
The Advanced search page allows a limited set of six parameters that
can be combined to create a search rule. For the purposes of storage
in the keyword table for recall, views, and to provide a title for
search rules, the field values are packed into a colon separated
string:

title:subtitle:description:category_type:category:station

MLB:Giants::::             San Francisco baseball games
:New York:::Sports event:  games involving any team from NY
:::movie:Action:HBO        action flix on Home Box Office

These strings are created and edited in a popup form on the proglist
plPowerSearch page. The strings are expanded to the SQL clauses for
proglist searches and for the SQL embedded in the kPowerSearch
rules. The Power Search rules should be roughly as fast as the other
search rules. In fact, the "MLB:Giants::::" rule may run faster, and
give more focused results (no ESPN Classics) than a Keyword search for
"San Francisco Giants".

Note that the interface only allows entering values for these six
fields and each non-empty field becomes an AND clause appended to
WHERE.  However, kPowerSearch (search type 1) rules allow any valid
SQL which the user may hand edit into the 'record' table (good luck =).
The 'subtitle' field may contain extensions to FROM. Other tables
(must begin with a comma) or JOINs for tables that would be needed for
the WHERE clauses would go here. The 'description' field is appended
to the WHERE clause. Parenthesis, AND, OR, NOT, etc. can be used to
build any arbitrary search rule.


git-svn-id: http://svn.mythtv.org/trunk@4568 4b91aca1-2a1f-0410-a73a-a78f2df169e8