Audio track scanning/selection tweak from danielk:
"I eliminated scanAudioStreams, and put that functionality in the general
ScanStreams method. Now it only adds an audio stream to the list if it
finds the codec for it in avlib. I've also changed ScanStreams so if it
finds an unidentified stream it keeps scanning the rest of the streams.
Finally, I added a check in GetFrame that calls SelectAudioTrack() if
the currently selected audio track is outside the bounds of our audio
stream list."


git-svn-id: http://svn.mythtv.org/trunk@4618 4b91aca1-2a1f-0410-a73a-a78f2df169e8