various misc fixes from IvanK <chepati(a)yahoo>
I took a look at mythweather and it appears that it expects the value of
SIUnuts ot be either YES or NO.  In my patch from yesterday I assumed Yes
would be ok.  I've attached a patch to change the Yes to YES.  Apply if you
think it's necessary.

The second patch is for three files that produce call-by-reference warnings.

The third patch takes care of an object of type Recording being used as array
even when it's not.

The last patch is for config/config.php.


git-svn-id: http://svn.mythtv.org/trunk@4726 4b91aca1-2a1f-0410-a73a-a78f2df169e8