Allow fonts to be declared inside the theme file...  Fonts declared inside the theme file go into a "global font map" and can be referenced from inside the various -ui files.
When checking for which font to use, the local font map specific to the window will be checked first, then the global map.  This allows theme designers to override the global font definition on a per window basis if desired.

These changes make it much easier for theme designers to create a consistent look and feel across the windows,  make large scale font changes as well as cut down on the clutter in the xml files.


git-svn-id: http://svn.mythtv.org/trunk@4778 4b91aca1-2a1f-0410-a73a-a78f2df169e8