Added two new recording types, kFindDailyRecord and kFindWeeklyRecord. The new recording types are intended for daily and weekly programs which meet the following criteria.
  They usually don't have unique programids or descriptions.

  A "new" episode is shown at (or about) the same time for daily
  programs or on the same day of the week and time for weekly
  programs.

  The new episode is repeated one or more times throughout the next 24
  hours for daily programs or seven days for weekly programs.

Because there are no unique programids or descriptions, Myth can't
tell the multiple showings apart.  For these new recording types, Myth
will automatically assign an artificial identifier to group together
all shows in the defined interval.  Myth will then try to record one
showing in each interval given other scheduling constraints.

The day of the week and time of the "new" episode are currently set
when the recording rule is created and can't be changed.  Suggestions
on how to make this changeable in the SR dialog are welcome.


git-svn-id: http://svn.mythtv.org/trunk@5090 4b91aca1-2a1f-0410-a73a-a78f2df169e8