Jeff Thompson <jthomps*gmail*com>
Hi, here's a small patch for mythweb's settings_channel.php. When
saving the channel info back to the db, it was using a non-existant
"videofilter" variable, causing videofilters info to be removed from
the db.

This patch changes the variable name to the correct value of
"videofilters" when performing the update.


git-svn-id: http://svn.mythtv.org/trunk@5125 4b91aca1-2a1f-0410-a73a-a78f2df169e8