Split the reschedule functionality into two parts.  The "match" part, which typically takes the most time, is now doen incrementally for only the recordids that are affected.  The intermediate results are stored in the recordmatch table and are used by the "placement" part which runs whenever anything changes.
This should greatly improve the responsiveness of the scheduler in
most cases when large numbers or record rules, particularly search
rules, are used.  Note there can still be a noticable to significant
delay at initial startup and when the program table is updated
(mythfilldatabase).  Those cases will be addressed in the future.


git-svn-id: http://svn.mythtv.org/trunk@5213 4b91aca1-2a1f-0410-a73a-a78f2df169e8