The last of the files.
With checkin the new theme "Minimalist-wide" is now usable.  This theme takes advantage of the new wide screen mode to put more information on the screen.

It's not 100% complete, it needs tweaking and there's likely a window or two I missed.

Screenshots are available at http://www.jdonavan.net/myth/wide_mode


git-svn-id: http://svn.mythtv.org/trunk@5526 4b91aca1-2a1f-0410-a73a-a78f2df169e8