Improved manual recording support.  Manual recordings are now preserved across mythfilldatabase runs.  Recurring daily and weekly manual programs are now supported.  GUI work was done by bjm.
Tweaked the overlap handling in the scheduler.  If all factors except
recordids are equal, prefer the program with the higher findid.  This
allows multiple FindWeekly rules to work reasonably for programs that
are shown semi-weekly.  Actually creating multiple FindWeekly rules
for the same program is left as an exercise for the user.


git-svn-id: http://svn.mythtv.org/trunk@5656 4b91aca1-2a1f-0410-a73a-a78f2df169e8