This is an unstable commit -- I'm doing this (and not packaging a version on my website) so that people can poke around with the new CLI code.
Anyway, this adds two major things -- first, I got multipass xvid encoding
working with ffmpeg.  Secondly, I've completely rewritten the cli-handling
code, and added support for an nuvexportrc file.  It's not yet complete,
since it really only completely works with the xvid exporters, but it's
enough to give people an idea of what they can do with it.


git-svn-id: http://svn.mythtv.org/trunk@11961 4b91aca1-2a1f-0410-a73a-a78f2df169e8