Refactored the video scan code.
If a file that exists in the database is missing when we do the scan, prompt the user before removing it from the DB.

Allow for multiple source directories, since the full path is stored anyway.  Use the "dir1:dir2"dir3" syntax for the video startup dir to take advantage of this.  When the video tree is in file browse mode, each base directory is a top level node instead of the generic "videos"

Remove ALL metadata when we remove a file from the database due to a failed scan, not just the videometadata table data.


git-svn-id: http://svn.mythtv.org/trunk@5738 4b91aca1-2a1f-0410-a73a-a78f2df169e8