Added options to Grid/Vertical/HorizontalConfigurationGroup to set extra padding between widgets to a minimum. This should be useful for cascading layout groups.
Previous default was to set the spacing (child <-> child) to equal margin
(border <-> child). Margin is set big for titles so spacing was sometimes
eating a lot of extra space.

The options zeroMargin and zeroSpace use a one pixel border
if set. zeroMargin should only be set if the Group does
not have a title.

Tweaked the Video Mode Settings screen with new options.


git-svn-id: http://svn.mythtv.org/trunk@6000 4b91aca1-2a1f-0410-a73a-a78f2df169e8