This adds some work-around code for the fluxbox window manager. You need to uncomment an appropriately named define at the top of tv_play for the hack to take effect.
I also added a DB query for the GUI Resolution, if it is set to zero width
and zero height, we assume the user does not want us to respect the
"GuiSizeForTV" property, but instead play the video at fullscreen.
Fullscreen resolution in the GUI and fullscreen resolution in the
playback window may differ if XRandR is being used. It used to be that
using XRandR always triggerred fullscreen playback, but with refresh
rate only changes possible now, this may not always be what the user
wants.


git-svn-id: http://svn.mythtv.org/trunk@6060 4b91aca1-2a1f-0410-a73a-a78f2df169e8