This removes the next to last exit() call in the myth libraries. It removes the exit() call in ThreadedFileWriter when an output file can not be opened (due to permissions problems, space, etc.)
If the file can not be opened the RingBuffer::IsOpen() call will
return false. No change here actually, except we don't no longer
exit the backend before this check could be made. But this does
that we need to check IsOpen() in NVR, and that "SETUP_RING_BUFFER"
can fail, so we now have it return false on failure. Finally, this
means we need to check the result in tv_rec.cpp, and tv_play.cpp.

In the old code the backend would go away leaving the frontend waiting
for a response to the SETUP_RING_BUFFER query.

This isn't a a big change, but I thought I'd explain why
MYTH_PROTO_VERSION had to change.

The only remaining exit() call in libmyth and libmythtv is in
ThemedMenuPrivate::handleAction(). I'm not sure how one would
go about fixing it without major changes. There are also some
left in mythui code, but I'm assuming there will be a larger
sync with the current MythTV tree to the new UI code.


git-svn-id: http://svn.mythtv.org/trunk@6284 4b91aca1-2a1f-0410-a73a-a78f2df169e8